Added new desktop file, and an --editor option.
If I want to launch the program from dmenu, or nautilus, it's easier
to have a --editor option. The precedence is

    --editor > $EDITOR > nano

Also added a xattrvi.desktop. If you have a reasonable file manager, now
you can simply open a file with 'Open With', choose xattrvi and edit the extended attributes.